src/requirements.txt
Flask==2.2.2
Flask-Cors==3.1.1
Flask-Bcrypt==1.0.0
Flask-JWT-Extended==4.4.4
mysql-connector-python==8.0.31
werkzeug==2.2.2
